requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-admin-honeypot==1.1.0
django-ckeditor==5.9.0
django-honeypot==0.9.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
idna==2.10
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
